:bug: maybe fix #5? (unreproducible)
from the stacks it looks the right editor goes away, at the same time an update is posted to the left editor, but before the onDidDestroy for the right editor is called? or maybe the right editor is tearing down and hasn't triggered onDidDestroy yet. I added a guard to circumvent the first case.  Will need reporter to test.  If it's still happening it's the second case which is not going to be easy to detect - I would need to reach into the editor and use and undocumented member -  editor.displayBuffer.isDestroyed() before calling SplitDiff.updateDiff