Load Git file data only when timeline is visible
This makes viewing and switching between files quicker when the
timeline is not visible. The only small downside would be that the
loading would happen when the timeline is toggled, rather than before
that. I think that this trade-off is optimal though. Merging this commit
to master would resolve #41.

This commit also makes a small change to only re-render the view if the
new active editor is not the same as the current active editor, slightly
optimizing performance when switching between the diff panes.