Upgrade split-diff package, fix blob render error
The split-diff package has several new features since the current
version (0.5.2), namely highlighting word changes within a line,
gutter line number highlighting, and several bug fixes. Package settings
for split-diff have to be manually set because the package is never
actually being activated. I also removed the manual scroll syncing
because I didn't see a problem with it in the new version. I could just
be missing it...

On another note, when there is no "linesAdded" or "linesRemoved"
properties in the git data, an error is thrown in the console. I made a
small change to use 0 in that case (because undefined + undefined ==
NaN...d3 throws an error for that). This would fix issue #42.